Куракин Матвей. 3822Б1ФИ1. OpenMP Naive Matrix Multiplication. Update

+++ 3822B1FI1/3_naive_gemm_omp/kurakin_matvey/naive_gemm_omp.cpp
@@ -6,9 +6,9 @@ std::vector<float> NaiveGemmOMP(const std::vector<float> &a,
                                 int n) {
     std::vector<float> res(n * n);
 #pragma omp parallel for
-    for(size_t i = 0; i < n; ++i){
-        for(size_t k = 0; k < n; ++k){
-            for(size_t j = 0; j < n; ++j){
+    for(int i = 0; i < n; ++i){
+        for(int k = 0; k < n; ++k){
+            for(int j = 0; j < n; ++j){
                 res[i * n + j] += a[i * n + k] * b[k * n + j];
             }
         }
